Removed integrationtest and moved shouldfallbackafterretries test into RecommendationServiceTest
